Restructured files for better clarity
