+++ presqt/api_v1/serializers/resource.py
@@ -11,14 +11,6 @@ class ResourcesSerializer(serializers.Serializer):
     title = serializers.CharField(max_length=256)
 
 
-class HashSerializer(serializers.Serializer):
-    """
-    Serializer for a hash object.
-    """
-    md5 = serializers.CharField(max_length=256)
-    sha256 = serializers.CharField(max_length=256)
-
-
 class ResourceSerializer(serializers.Serializer):
     """
     Serializer for a give resource object.
@@ -30,5 +22,5 @@ class ResourceSerializer(serializers.Serializer):
     date_created = serializers.DateField()
     date_modified = serializers.DateField()
     size = serializers.IntegerField()
-    hashes = HashSerializer()
+    hashes = serializers.DictField()
     extra = serializers.DictField()
