+++ presqt/api_v1/utilities/io/write_file.py
@@ -1,6 +1,5 @@
 import json
 import os
-import zipfile
 
 
 def write_file(file_path, contents, is_json=False):
@@ -23,20 +22,3 @@ def write_file(file_path, contents, is_json=False):
     else:
         with open(file_path, 'wb') as outfile:
             outfile.write(contents)
-
-def zip_directory(destination_path, source_path):
-    """
-    Zip a directory to a specified path.
-
-    Parameters
-    ----------
-    destination_path : str
-        Path to save the zip file
-    source_path :
-        Path of the directory to be zipped.
-    """
-    my_zip_file = zipfile.ZipFile(destination_path, "w")
-    for root, dirs, files in os.walk(source_path):
-        for file in files:
-            my_zip_file.write(os.path.join(root, file))
-    my_zip_file.close()
