+++ presqt/api_v1/utilities/utils/function_router.py
@@ -19,6 +19,14 @@ class FunctionRouter(object):
         {target_name}_download
 
     """
+    @classmethod
+    def get_function(cls, target_name, action):
+        """
+        Extracts the getattr() function call to this class method so the code using this class
+        is easier to work with.
+        """
+        return getattr(cls, '{}_{}'.format(target_name, action))
+
     osf_resource_collection = osf_fetch_resources
     osf_resource_detail = osf_fetch_resource
     osf_resource_download = osf_download_resource
